+++ README.md
@@ -1,5 +1,3 @@
 # Prophetic Strategies
 
 A simple RAG app using Lacan's Seminar Book 5. Access here.
-
-Uses openai, streamlit, sqlite, pandas.
