+++ src/prophetic-strategies/streamlit_app.py
@@ -1,8 +1,7 @@
 import streamlit as st
 from pathlib import Path
-
 from utils.utils import fake_stream
-from utils.utils import rag
+from utils.semantic_search import search
 
 STRATEGIES = {
     # "Navigator": Path("db/oblique.db"),
@@ -12,9 +11,10 @@ STRATEGIES = {
 with st.sidebar:
     st.header("Parameters")
     strategy_name = st.selectbox("Strategy", list(STRATEGIES.keys()))
-    strategy = STRATEGIES[strategy_name]  # type: ignore
+    db = STRATEGIES[strategy_name]  # type: ignore
 
 st.title("Prophetic Strategies")
+# st.caption("[Source](https://github.com/noah-art3mis/prophetic-strategies)")
 st.write("")
 question = st.text_input("What is it that you desire (not) to know?")
 st.write("")
@@ -22,7 +22,7 @@ st.write("")
 
 
 if question != "":
-    result = rag(question, strategy)
+    result = search(question, db)
     st.write_stream(fake_stream(result["content"]))
     st.write("")
     st.caption(result["book"])
